requirements.txt
# ── Scraping & Parsing ──────────────────────────────────────────
httpx>=0.27
mwparserfromhell>=0.7

# ── Embedding & Search ─────────────────────────────────────────
chromadb>=1.0
numpy>=2.0
rank-bm25>=0.2
FlashRank>=0.2

# ── RAG / LLM & TTS & Image Gen ───────────────────────────────
google-genai>=1.0
python-dotenv>=1.0

# ── Subtitles ──────────────────────────────────────────────────
faster-whisper>=1.0

# ── Video Assembly ─────────────────────────────────────────────
moviepy>=2.0
Pillow>=10.0

# ── Web App ────────────────────────────────────────────────────
Flask>=3.0

# ── CLI helpers ────────────────────────────────────────────────
tqdm>=4.60
